+++ foundrytools_cli_2/cli/commands/converter/__init__.py
@@ -1,0 +1,3 @@
+from foundrytools_cli_2.cli.commands.converter.cli import cli
+
+__all__ = ["cli"]
