+++ foundrytools_cli_2/cli/commands/name.py
