+++ foundrytools_cli_2/cli/commands/utils/__init__.py
@@ -1,0 +1,3 @@
+from foundrytools_cli_2.cli.commands.utils.cli import cli
+
+__all__ = ["cli"]
