--- foundrytools_cli_2/cli/otf/__init__.py
@@ -1,3 +1,0 @@
-from foundrytools_cli_2.cli.otf.cli import cli
-
-__all__ = ["cli"]
