--- foundrytools_cli_2/cli/ttf/__init__.py
@@ -1,3 +1,0 @@
-from foundrytools_cli_2.cli.ttf.cli import cli
-
-__all__ = ["cli"]
